fix type

+++ include/ivf/ivf_flat.h
@@ -9,12 +9,12 @@
 #include <omp.h>
 
 template<typename T>
-void ivf_flat_insert(int32_t nx, const T* x_in, int32_t dim,
-                     int32_t nlist, const float* centroids,
+void ivf_flat_insert(int64_t nx, const T* x_in, int64_t dim,
+                     int64_t nlist, const float* centroids,
                      std::vector<std::vector<T>> &codes,
                      std::vector<std::vector<int32_t>> &ids) {
 
-    std::unique_ptr<int32_t []> assign(new int32_t[nx]);
+    std::unique_ptr<int64_t []> assign(new int64_t[nx]);
     std::unique_ptr<float []> dis(new float[nx]);
     elkan_L2_assign(x_in, centroids, dim, nx, nlist, assign.get(), dis.get());
     dis = nullptr;
@@ -24,15 +24,15 @@ void ivf_flat_insert(int32_t nx, const T* x_in, int32_t dim,
 
 #pragma omp parallel
     {
-        int nt = omp_get_num_threads();
-        int rank = omp_get_thread_num();
+        int64_t nt = omp_get_num_threads();
+        int64_t rank = omp_get_thread_num();
 
         // this thread is taking care of centroids c0:c1
-        size_t c0 = (nlist * rank) / nt;
-        size_t c1 = (nlist * (rank + 1)) / nt;
+        int64_t c0 = (nlist * rank) / nt;
+        int64_t c1 = (nlist * (rank + 1)) / nt;
 
-        for (int32_t i = 0; i < nx; i++) {
-            int32_t ci = assign[i];
+        for (int64_t i = 0; i < nx; i++) {
+            int64_t ci = assign[i];
             if (ci >= c0 && ci < c1) {
                 codes[ci].resize(codes[ci].size() + dim);
                 memcpy(codes[ci].data() + ids[ci].size() * dim,
@@ -41,15 +41,14 @@ void ivf_flat_insert(int32_t nx, const T* x_in, int32_t dim,
             }
         }
     }
-
 }
 
 template<class C, typename T, typename R>
-void ivf_flat_search(int32_t nq, const T* q_in, int32_t dim,
-                     int32_t nlist, const float* centroids,
+void ivf_flat_search(int64_t nq, const T* q_in, int64_t dim,
+                     int64_t nlist, const float* centroids,
                      std::vector<std::vector<T>> &codes,
                      std::vector<std::vector<int32_t>> &ids,
-                     int32_t nprobe, int32_t topk,
+                     int64_t nprobe, int64_t topk,
                      typename C::T * value,
                      typename C::TI * labels,
                      Computer<T, T, typename C::T> comptuer) {
@@ -61,7 +60,7 @@ void ivf_flat_search(int32_t nq, const T* q_in, int32_t dim,
     coarse_dis = nullptr;
 
 #pragma omp parallel for
-    for (int32_t i = 0; i < nq; i++) {
+    for (int64_t i = 0; i < nq; i++) {
         auto *q_i = q_in + i * dim;
 
         auto * __restrict val_ = value  + i * topk;
@@ -70,13 +69,13 @@ void ivf_flat_search(int32_t nq, const T* q_in, int32_t dim,
         // init heap
         heap_heapify<C>(topk, val_, ids_);
 
-        for (int32_t j = 0; j < nprobe; j++) {
+        for (int64_t j = 0; j < nprobe; j++) {
             auto idx_j = idx[i * nprobe + j];
             auto code = codes[idx_j].data();
             auto id = ids[idx_j].data();
-            int32_t count = ids[idx_j].size();
+            int64_t count = ids[idx_j].size();
 
-            for (int32_t k = 0; k < count; k++) {
+            for (int64_t k = 0; k < count; k++) {
                 auto disij = comptuer (q_i, code, dim);
                 if (C::cmp(val_[0], disij)) {
                     heap_swap_top<C>(topk, val_, ids_, disij, id[k]);
